sync

+++ minference/ops/block_sparse_flash_attention.py
@@ -348,9 +348,9 @@ def _triton_block_sparse_attn_fwd_kernel_with_kvcache(
         # q_absolute = q_preced_len + start_m * BLOCK_M
         # abs_offs_m = q_absolute + tl.arange(0, BLOCK_M)
         # * abs_offs_m = q_preced_len + start_m * BLOCK_M + tl.arange(0, BLOCK_M)
-        abs_offs_m = q_preced_len + offs_m
+        abs_offs_m = (q_preced_len + offs_m).to(tl.int32)
         
-        cols = start_n + offs_n
+        cols = (start_n + offs_n).to(tl.int32)
         causal_mask = cols[None, :] <= abs_offs_m[:, None]
 
         # * qk \in (BLOCK_M, BLOCK_DMODEL) \times (BLOCK_DMODEL, BLOCK_N) -> (BLOCK_M, BLOCK_N)
